requirements.txt
google-cloud-vision
python-dotenv
requests
